Started refactoring into separate failes and added input CSV file input
