+++ .gitignore
@@ -1,3 +1,2 @@
 /target
-/test_data
-.DS_Stoore
+.DS_Store
