+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "qcma"
+name = "qmca"
 version = "0.1.0"
 authors = ["mtseng15 <tseng.micah@gmail.com>"]
 edition = "2018"
@@ -7,7 +7,8 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-structopt = { version = "0.3", default-features = false }
+# structopt = { version = "0.3", default-features = false }
+clap = "2.33.1"
 failure = "0.1.8"
 exitfailure = "0.5.1"
 csv = "1.1"
